nginx: add `reachability_healthcheck`
this healthcheck tests all the web services reachability from the docker
network `http://<service>` and for external server access
`https://<service>.$KITSPACE_DOMAIN` in prod/staging envs.
If one of the services is unreachable it will reload nginx. If reloading
nginx fails it exits marks the container as unhealthy.
The healthcheck only doesn't restart the container if it's unhealthy
though. Restarting unhealthy containers should be done by the host
machine.